feat: 장바구니 항목 그룹화 구현
